naming and move `ComposedPhysicalExtensionCodec` to datafusion
